# models/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@yourorganization.com
    name: SSM API Support
  description: |
    API for secure cryptographic key management using PKCS#11 and HSM.

    SSM provides secure operations for:
    - AES, DES, DES3 key generation
    - Data encryption and decryption
    - Key storage and management
    - HSM/SoftHSM integration

    ## Authentication
    The API works through Unix Domain Sockets for enhanced security.
    Also supports HTTPS with TLS certificates.

    ## Data Formats
    - All binary data (plaintext, ciphertext, IV) should be in Base64/Hex
    - Responses include timestamps in RFC3339 format
    - Errors follow RFC 7807 standard (Problem Details)
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: SSM (Secure Storage Manager) API
  version: 1.0.0
...
servers:
- description: Local Unix Socket Server
  url: http://localhost
- description: Local HTTPS Server
  url: https://localhost:9001
...
paths:
  /generate-aes-key:
    post:
      description: |
        Generates a new symmetric AES key in the HSM with the specified size.
        The key is stored persistently and referenced by its label.
      operationId: generateAESKey
      requestBody:
        content:
          application/json:
            examples:
              aes256:
                summary: AES-256 Key
                value:
                  id: 1
                  bits: 256
              aes128:
                summary: AES-128 Key
                value:
                  id: 2
                  bits: 128
            schema:
              $ref: '#/components/schemas/GenAESKeyRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              examples:
                success:
                  summary: Successful response
                  value:
                    handle: 123456789
                    id: 2
                    bits: 256
              schema:
                $ref: '#/components/schemas/GenAESKeyResponse'
          description: Key generated successfully
        "400":
          content:
            application/json:
              examples:
                validation:
                  summary: Validation Error
                  value:
                    title: Validation Error
                    detail: Label is required
                    status: 400
                    error: validation_failed
                badJson:
                  summary: Invalid JSON
                  value:
                    title: Invalid JSON
                    detail: Failed to parse request body
                    status: 400
                    error: bad_json
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Invalid request
        "405":
          content:
            application/json:
              example:
                title: Method Not Allowed
                detail: Only POST method is allowed
                status: 405
                error: method_not_allowed
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: HTTP method not allowed
        "500":
          content:
            application/json:
              examples:
                hsm_error:
                  summary: HSM Error
                  value:
                    title: Key Generation Failed
                    detail: "Failed to generate AES key: pkcs11 session invalid"
                    status: 500
                    error: hsm_error
                general:
                  summary: General Error
                  value:
                    title: Internal Server Error
                    detail: An unexpected error occurred
                    status: 500
                    error: internal_error
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal server error
      summary: Generate new AES key
      tags:
      - Key Management
  /generate-des-key:
    post:
      description: |
        Generates a new DES key in the HSM.
        The key is stored persistently and referenced by its label.
      operationId: generateDESKey
      requestBody:
        content:
          application/json:
            examples:
              simple:
                summary: DES Key
                value:
                  id: 1
            schema:
              $ref: '#/components/schemas/GenDESKeyRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenDESKeyResponse'
          description: DES key generated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal server error
      summary: Generate new DES key
      tags:
      - Key Management
  /generate-des3-key:
    post:
      description: |
        Generates a new DES3 key in the HSM.
        The key is stored persistently and referenced by its label.
      operationId: generateDES3Key
      requestBody:
        content:
          application/json:
            examples:
              simple:
                summary: DES3 Key
                value:
                  id: 1
            schema:
              $ref: '#/components/schemas/GenDES3KeyRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenDES3KeyResponse'
          description: DES3 key generated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal server error
      summary: Generate new DES3 key
      tags:
      - Key Management
...
  /store-key:
    delete:
      description: |
        Deletes a cryptographic key from the HSM.
        The key is identified by its label and optional ID.
      operationId: deleteKey
      requestBody:
        content:
          application/json:
            examples:
              simple:
                summary: Delete key
                value:
                  key_label: ImportedKey
                  id: 1
            schema:
              $ref: '#/components/schemas/DeleteKeyRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteKeyResponse'
          description: Key deleted successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Key not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal server error
      summary: Delete key
      tags:
      - Key Management
    post:
      description: |
        Stores an existing cryptographic key in the HSM.
        The key must be provided in the appropriate format.
      operationId: storeKey
      requestBody:
        content:
          application/json:
            examples:
              simple:
                summary: Store key
                value:
                  key_label: ImportedKey
                  id: 3
                  key_value: YmFzZTY0X2VuY29kZWRfa2V5
                  key_type: AES
            schema:
              $ref: '#/components/schemas/StoreKeyRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              examples:
                success:
                  summary: Key stored
                  value:
                    handle: 987654321
                    cipher_key: encrypted_key_value
              schema:
                $ref: '#/components/schemas/StoreKeyResponse'
          description: Key stored successfully
        "400":
          content:
            application/json:
              examples:
                validation:
                  summary: Validation error
                  value:
                    title: Validation Error
                    detail: Label is required
                    status: 400
                    error: validation_failed
                badJson:
                  summary: Invalid JSON
                  value:
                    title: Invalid JSON
                    detail: Failed to parse request body
                    status: 400
                    error: bad_json
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Invalid request
        "500":
          content:
            application/json:
              examples:
                hsm_error:
                  summary: HSM error
                  value:
                    title: Key Generation Failed
                    detail: "Failed to generate AES key: pkcs11 session invalid"
                    status: 500
                    error: hsm_error
                general:
                  summary: General error
                  value:
                    title: Internal Server Error
                    detail: An unexpected error occurred
                    status: 500
                    error: internal_error
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal server error
      summary: Store existing key
      tags:
      - Key Management
    put:
      description: |
        Updates an existing cryptographic key in the HSM.
        Replaces the key value with a new one while keeping the same label and ID.
      operationId: updateKey
      requestBody:
        content:
          application/json:
            examples:
              simple:
                summary: Update key example
                value:
                  key_label: ImportedKey
                  id: 1
                  key_value: bmV3X2tleV92YWx1ZQ==
                  key_type: AES
            schema:
              $ref: '#/components/schemas/UpdateKeyRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateKeyResponse'
          description: Key updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Invalid request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Key not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal server error
      summary: Update key
      tags:
      - Key Management
...
components:
...
  schemas:
    GenAESKeyRequest:
      example:
        id: key001
        bits: 256
      properties:
        id:
          description: Unique key identifier
          type: integer
        bits:
          description: Key size in bits
          enum:
          - 128
          - 192
          - 256
          example: 256
          type: integer
      required:
      - bits
      - id
      type: object
    GenAESKeyResponse:
      example:
        bits: 256
        handle: 123456789
        id: 2
      properties:
        handle:
          description: HSM key handle
          example: 123456789
          format: uint
          type: integer
        id:
          description: Generated key identifier
          example: 2
          type: integer
        bits:
          description: Size of the generated key in bits
          example: 256
          type: integer
      type: object
...
    StoreKeyRequest:
      example:
        key_label: ImportedKey
        id: 2
        key_value: abcde1231312
        key_type: AES
      properties:
        key_label:
          description: Label for the stored key
          example: ImportedKey
          minLength: 1
          type: string
        id:
          description: Unique identifier
          example: 2
          type: integer
        key_value:
          description: Key value in Base64
          example: !!binary |-
            WVdKalpHVXhNak14TXpFeQ==
          format: byte
          type: string
        key_type:
          description: Type of cryptographic key
          enum:
          - AES
          - DES
          - DES3
          example: AES
          type: string
      required:
      - id
      - key_label
      - key_type
      - key_value
      type: object
    StoreKeyResponse:
      example:
        handle: 987654321
        cipher_key: encrypted_key_value
      properties:
        handle:
          description: Handle of the stored key
          example: 987654321
          format: uint
          type: integer
        cipher_key:
          description: Stored encrypted key
          example: encrypted_key_value
          type: string
      type: object
    ProblemDetails:
      description: Detalles de error según RFC 7807
      example:
        instance: /generate-aes-key
        detail: Label is required
        title: Validation Error
        error: validation_failed
        status: 400
      properties:
        title:
          description: Short title of the problem
          example: Validation Error
          type: string
        detail:
          description: Detailed description of the problem
          example: Label is required
          type: string
        status:
          description: HTTP status code
          example: 400
          type: integer
        error:
          description: Internal error code
          example: validation_failed
          type: string
        instance:
          description: URI identifying the specific occurrence
          example: /generate-aes-key
          format: uri
          type: string
      type: object
    GenDESKeyRequest:
      example:
        id: 1
      properties:
        id:
          description: Unique key identifier
          example: 1
          type: integer
      required:
      - id
      type: object
    GenDESKeyResponse:
      example:
        handle: 123456789
        id: 2
      properties:
        handle:
          description: HSM key handle
          example: 123456789
          type: integer
        id:
          description: Key identifier
          example: 3
          type: integer
      type: object
    GenDES3KeyRequest:
      example:
        id: 1
      properties:
        id:
          description: Unique key identifier
          example: 1
          type: integer
      required:
      - id
      type: object
    GenDES3KeyResponse:
      example:
        handle: 123456789
        id: 1
      properties:
        handle:
          description: HSM key handle
          example: 123456789
          type: integer
        id:
          description: Key identifier
          example: 1
          type: integer
      type: object
    DeleteKeyRequest:
      example:
        key_label: ImportedKey
        id: 1
      properties:
        key_label:
          description: Label of the key to delete
          example: ImportedKey
          maxLength: 256
          minLength: 1
          type: string
        id:
          description: Key identifier (optional)
          example: 1
          maxLength: 64
          type: integer
      required:
      - id
      - key_label
      type: object
    DeleteKeyResponse:
      example:
        message: Key deleted successfully
        key_label: ImportedKey
      properties:
        message:
          description: Confirmation message
          example: Key deleted successfully
          type: string
        key_label:
          description: Label of the deleted key
          example: ImportedKey
          type: string
      required:
      - key_label
      - message
      type: object
    UpdateKeyRequest:
      example:
        key_label: ImportedKey
        id: 2
        key_value: 1213afabe213be2
        key_type: AES
      properties:
        key_label:
          description: Label of the key to update
          example: ImportedKey
          maxLength: 256
          minLength: 1
          type: string
        id:
          description: Key identifier
          example: 2
          type: integer
        key_value:
          description: New key value in hexadecimal format
          example: bmV3X2tleV92YWx1ZQ==
          type: string
        key_type:
          description: Type of cryptographic key
          enum:
          - AES
          - DES
          - DES3
          example: AES
          type: string
      required:
      - id
      - key_label
      - key_type
      - key_value
      type: object
    UpdateKeyResponse:
      example:
        message: Key updated successfully
        handle: 987654321
        key_label: ImportedKey
        cipher_key: encrypted_key_value
      properties:
        message:
          description: Confirmation message
          example: Key updated successfully
          type: string
        handle:
          description: New HSM key handle
          example: 987654321
          type: integer
        key_label:
          description: Label of the updated key
          example: ImportedKey
          type: string
        cipher_key:
          description: Encrypted key value (optional)
          example: encrypted_key_value
          type: string
      required:
      - handle
      - key_label
      - message
      type: object
...